restructure officer listing component to allow either array of officers as input or a semester
