+++ frontend-site/src/store/officers.js
@@ -1,6 +1,5 @@
 import lugApi from '@/modules/LugApi';
 
-// TODO: decide to keep or drop; only needed for caching, but current implementation doesn't need it
 export default {
   namespaced: true,
   state: {
@@ -13,6 +12,22 @@ export default {
         throw new Error(`New data is not an array: ${JSON.stringify(data)}`);
       }
       state.data = data;
+    },
+  },
+  getters: {
+    currentSemester () {
+      const currentDate = new Date();
+      const currentMonth = currentDate.getMonth() + 1; // getMonth is 0 indexed, so add 1
+      const currentYear = currentDate.getUTCFullYear();
+      let currentSemester = '';
+      if (currentMonth <= 5) { // up to May
+        currentSemester = 'SPRING';
+      } else if (currentMonth <= 8) { // up to August
+        currentSemester = 'SUMMER';
+      } else {
+        currentSemester = 'FALL';
+      }
+      return `${currentSemester}_${currentYear}`;
     },
   },
   actions: {
